Initial implementation of OTP endpoints

+++ powerauth-nextstep-model/src/main/java/io/getlime/security/powerauth/lib/nextstep/model/response/GetOperationDetailResponse.java
@@ -58,7 +58,7 @@ public class GetOperationDetailResponse {
     private final List<AfsActionDetail> afsActions = new ArrayList<>();
     private OperationFormData formData;
     private AuthMethod chosenAuthMethod;
-    private Integer remainingAttempts;
+    private Long remainingAttempts;
     private ApplicationContext applicationContext;
 
     public boolean isExpired() {
